+++ back/src/__tests__/helpers/checkRoleAccess.test.ts
@@ -1,20 +1,20 @@
-import { USER_ROLE } from '@prisma/client'
 import { test, describe, expect } from 'vitest'
 import { ForbiddenError } from '../../helpers/errors'
 import { checkRoleAccess } from '../../helpers'
+import { UserRole } from '../../schemas/users/userSchema'
 
 describe('tests checkRoleAccess function', () => {
   test('should not throw error when user has required access level', () => {
-    const requiredRole = USER_ROLE.MENTOR
-    const userRole = USER_ROLE.ADMIN
+    const requiredRole = UserRole.MENTOR
+    const userRole = UserRole.ADMIN
     expect(() => {
       checkRoleAccess(requiredRole, userRole)
     }).not.toThrow()
   })
 
   test('should throw ForbiddenError when user does not have required access level', () => {
-    const requiredRole = USER_ROLE.MENTOR
-    const userRole = USER_ROLE.REGISTERED
+    const requiredRole = UserRole.MENTOR
+    const userRole = UserRole.REGISTERED
     expect(() => {
       checkRoleAccess(requiredRole, userRole)
     }).toThrow(ForbiddenError)