+++ back/src/controllers/auth/authMeController.ts
@@ -12,7 +12,6 @@ export const authMeController: Middleware = async (ctx: Context) => {
     select: {
       id: true,
       name: true,
-      role: true,
       status: true,
       avatarId: true,
     },