+++ packages/ui/src/__tests__/molecules/Tabs/Tabs.test.tsx
@@ -1,5 +1,6 @@
 import { render, screen, waitFor, fireEvent } from '@testing-library/react'
 import { Tabs } from '../../../components/molecules/Tabs'
+import { colors } from '../../../styles'
 
 const tabsTestData = [
   {
@@ -74,6 +75,37 @@ describe('TabsManager component', () => {
     )
     expect(
       screen.queryByText('Test component content B')
+    ).not.toBeInTheDocument()
+  })
+
+  it('shows default active tab when provided', () => {
+    render(<Tabs tabsData={tabsTestData} defaultActiveTab={1} />)
+
+    const tabA = screen.getByRole('button', {
+      name: /test a/i,
+    })
+
+    const tabB = screen.getByRole('button', {
+      name: /test b/i,
+    })
+
+    expect(tabA).toHaveStyle({ color: `${colors.gray.gray2}` })
+    expect(tabA).toHaveStyle({
+      borderBottom: `1px solid ${colors.primaryLight}`,
+    })
+
+    expect(tabB).toHaveStyle({ color: `${colors.black.black1}` })
+    expect(tabB).toHaveStyle({
+      borderBottom: `1px solid ${colors.primary}`,
+    })
+
+    expect(screen.getByText('Test component content B')).toBeInTheDocument()
+
+    expect(
+      screen.queryByText('Test component content A')
+    ).not.toBeInTheDocument()
+    expect(
+      screen.queryByText('Test component content C')
     ).not.toBeInTheDocument()
   })
 })