+++ packages/ui/src/components/molecules/Tabs/TabItem.tsx
@@ -1,43 +1,41 @@
 import styled from 'styled-components'
 import { colors, dimensions, font } from '../../../styles'
 
-//ERROR: isactive? https://stackoverflow.com/questions/76935768/react-does-not-recognize-the-isactive-prop-on-a-dom-element
-
-const TabItemLi = styled.li<{ isActive: boolean }>`
+const TabItemLi = styled.li<{ $isactive: boolean }>`
   list-style-type: none;
   margin: 0;
   align-self: baseline;
   padding: ${dimensions.spacing.xs} ${dimensions.spacing.xl};
-  cursor: ${(props) => (props.isActive ? 'default' : 'pointer')};
+  cursor: ${(props) => (props.$isactive ? 'default' : 'pointer')};
   color: ${(props) =>
-    props.isActive ? `${colors.black.black1}` : `${colors.gray.gray3}`};
+    props.$isactive ? `${colors.black.black1}` : `${colors.gray.gray3}`};
   font-weight: ${(props) =>
-    props.isActive ? `${font.bold}` : `${font.medium}`};
+    props.$isactive ? `${font.bold}` : `${font.medium}`};
   border-bottom: 1px solid
-    ${(props) => (props.isActive ? `${colors.primary}` : 'transparent')};
+    ${(props) => (props.$isactive ? `${colors.primary}` : 'transparent')};
   box-shadow: ${(props) =>
-    props.isActive ? `0px 2px 0px ${colors.primary}` : 'transparent'};
+    props.$isactive ? `0px 2px 0px ${colors.primary}` : 'transparent'};
   -webkit-box-shadow: 0px 2px 0px
-    ${(props) => (props.isActive ? `${colors.primary}` : 'transparent')};
+    ${(props) => (props.$isactive ? `${colors.primary}` : 'transparent')};
   -moz-box-shadow: 0px 2px 0px
-    ${(props) => (props.isActive ? `${colors.primary}` : 'transparent')};
+    ${(props) => (props.$isactive ? `${colors.primary}` : 'transparent')};
   font-family: ${font.fontFamily};
 
   &:hover {
     color: ${(props) =>
-      props.isActive ? `${colors.black.black1}` : `${colors.gray.gray2}`};
+      props.$isactive ? `${colors.black.black1}` : `${colors.gray.gray2}`};
     border-bottom: 1px solid
       ${(props) =>
-        props.isActive ? `${colors.primary}` : `${colors.primaryLight}`};
+        props.$isactive ? `${colors.primary}` : `${colors.primaryLight}`};
     -webkit-box-shadow: 0px 2px 0px
       ${(props) =>
-        props.isActive ? `${colors.primary}` : `${colors.primaryLight}`};
+        props.$isactive ? `${colors.primary}` : `${colors.primaryLight}`};
     -moz-box-shadow: 0px 2px 0px
       ${(props) =>
-        props.isActive ? `${colors.primary}` : `${colors.primaryLight}`};
+        props.$isactive ? `${colors.primary}` : `${colors.primaryLight}`};
     box-shadow: 0px 2px 0px
       ${(props) =>
-        props.isActive ? `${colors.primary}` : `${colors.primaryLight}`};
+        props.$isactive ? `${colors.primary}` : `${colors.primaryLight}`};
     font-weight: ${font.bold};
   }
 `
@@ -45,14 +43,14 @@ const TabItemLi = styled.li<{ isActive: boolean }>`
 type TTabItem = {
   title: string
   index: number
-  isActive: boolean
+  $isactive: boolean
   setActiveTab: (index: number) => void
 }
 
-const TabItem = ({ title, isActive, setActiveTab, index }: TTabItem) => (
+const TabItem = ({ title, $isactive, setActiveTab, index }: TTabItem) => (
   <TabItemLi
     onClick={() => setActiveTab(index)}
-    isActive={isActive}
+    $isactive={$isactive}
     role="button"
   >
     {title}