+++ packages/ui/src/stories/molecules/Tabs/Tabs.stories.tsx
@@ -1,30 +1,33 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Tabs from '../../../components/molecules/Tabs/Tabs'
-import { Button } from '../../../components/atoms'
+import { Tabs } from '../../../components/molecules/Tabs'
+import { Text } from '../../../components/atoms'
 
-const ContentA = () => <Button>The content in the Tab A</Button>
-const ContentB = () => <>The content in the Tab B</>
+const ContentA = () => <Text>The content in the Tab A</Text>
+const ContentB = () => <Text>The content in the Tab B</Text>
+const ContentBActive = () => (
+  <Text>The content in the Tab B. Tab active by default.</Text>
+)
+const ContentC = () => <Text>The content in the Tab C</Text>
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
   title: 'Molecules/Tabs',
   component: Tabs,
   parameters: {
     layout: 'centered',
   },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
+
   tags: ['autodocs'],
-  // More on argTypes: https://storybook.js.org/docs/api/argtypes
+
   argTypes: {
+    defaultActiveTab: {
+      control: {
+        type: 'number',
+        min: 0,
+        max: 2,
+      },
+    },
     tabsData: {
       control: 'object',
-      id: {
-        name: 'id',
-        type: 'string',
-        description: 'Unique identifier for the tab',
-        control: 'text',
-      },
-
       title: {
         name: 'title',
         type: 'string',
@@ -32,31 +35,12 @@ const meta = {
         required: true,
         control: 'text',
       },
-
       tabComponent: {
         name: 'tabComponent',
         type: 'object',
         description: 'React component to render for the tab',
         required: true,
         control: 'object',
-      },
-
-      requiredRole: {
-        control: 'object',
-        name: 'requiredRoles',
-        type: 'array',
-        description: 'List of roles required to view the tab',
-        oneOf: [
-          {
-            type: 'array',
-            items: {
-              type: 'string',
-            },
-          },
-          {
-            type: 'string',
-          },
-        ],
       },
     },
   },
@@ -65,21 +49,24 @@ const meta = {
 export default meta
 type Story = StoryObj<typeof meta>
 
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
 export const Default: Story = {
   args: {
     tabsData: [
       { title: 'Tab Item A', tabComponent: <ContentA /> },
       { title: 'Tab Item B', tabComponent: <ContentB /> },
+      { title: 'Tab Item C', tabComponent: <ContentC /> },
     ],
+    defaultActiveTab: 0,
   },
 }
 
-export const Active: Story = {
+export const SecondTabActive: Story = {
   args: {
     tabsData: [
       { title: 'Tab Item A', tabComponent: <ContentA /> },
-      { title: 'Tab Item B', tabComponent: <ContentB /> },
+      { title: 'Tab Item B', tabComponent: <ContentBActive /> },
+      { title: 'Tab Item C', tabComponent: <ContentC /> },
     ],
+    defaultActiveTab: 1,
   },
 }