+++ sso/src/openapi/components/responses.ts
@@ -106,7 +106,7 @@ export const getUserResponse = {
   description: 'Token is valid and user information is returned.',
   content: {
     'application/json': {
-      schema: userSchema.pick({ dni: true, email: true }),
+      schema: userSchema.pick({ dni: true, email: true, role: true }),
     },
   },
 }